Agregar alertas a metodos por medio de alertify

+++ src/app/pages/show/show.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { faPlus, faPen, faTrash } from '@fortawesome/free-solid-svg-icons';
 import { Product } from 'src/app/models/product.model';
+import { AlertifyService } from 'src/app/services/alertify.service';
 import { CrudService } from 'src/app/services/crud.service';
 
 @Component({
@@ -13,7 +14,8 @@ export class ShowComponent implements OnInit {
   products: Product[] = [];
 
   constructor(
-    private _crudService : CrudService
+    private _crudService : CrudService,
+    private _alertifyService: AlertifyService
   ) { }
 
   faPlus = faPlus;
@@ -28,10 +30,17 @@ export class ShowComponent implements OnInit {
   }
 
   delete(id:any, index:any){
-    this._crudService.deleteProduct(id)
-      .subscribe(() =>{
-        this.products.splice(index, 1);   // Elimina el producto de la vista
-      })
+
+    this._alertifyService.confirm({
+      message: 'Are you sure you want to delete this product?',
+      callback_delete: () => {
+        this._crudService.deleteProduct(id)
+          .subscribe(() =>{
+            this.products.splice(index, 1);   // Elimina el producto de la vista
+          })
+      }
+    })
+
   }
 
 }
